Design

+++ src/main/java/com/svalero/dao/ProductDAO.java
@@ -5,6 +5,7 @@ import org.jdbi.v3.sqlobject.statement.SqlQuery;
 import org.jdbi.v3.sqlobject.statement.SqlUpdate;
 import org.jdbi.v3.sqlobject.statement.UseRowMapper;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface ProductDAO {
@@ -14,7 +15,7 @@ public interface ProductDAO {
     List<Product> getProducts();
 
     @SqlUpdate("INSERT INTO products (name, description, price, image) VALUES (?, ?, ?, ?)")
-    void addProduct(String name, String description, Double price, String image);
+    void addProduct(String name, String description, BigDecimal price, String image);
 
     @SqlQuery("SELECT * FROM products WHERE product_id = ?")
     @UseRowMapper(ProductMapper.class)
@@ -24,6 +25,6 @@ public interface ProductDAO {
     void removeProduct(int product_id);
 
     @SqlUpdate("UPDATE products SET name = ?, description = ?, price = ?, image = ? WHERE product_id = ?")
-    void editProduct(String name, String description, Double price, String image, int product_id);
+    void editProduct(String name, String description, BigDecimal price, String image, int product_id);
 
 }
